fix: fixed download percentage not moving
